# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "212messenger App API Doc",
    "description": "API Documentation for 212messenger App",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
  },
  {
      "url": "http://18.130.137.253//api"
  }
  ],
  "paths": {
    "/v1/auth/send-otp": {
  "post": {
    "tags": ["Authentication"],
    "summary": "Send OTP API",
    "description": "API to send an OTP to the provided email address.",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "vishvadattfreshcode@gmail.com",
                "description": "Email address to which the OTP will be sent."
              }
            },
            "required": ["email"]
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "OTP sent successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 1,
                  "description": "Status of the request. 1 indicates success."
                },
                "message": {
                  "type": "string",
                  "example": "Otp Sent Successfully",
                  "description": "Response message indicating the result of the request."
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Bad Request - Invalid input or missing parameters."
      },
      "500": {
        "description": "Internal Server Error - An error occurred on the server."
      }
    }
  }
},
    "/v1/auth/verify-otp": {
  "post": {
    "tags": ["Authentication"],
    "summary": "Verify OTP API",
    "description": "API to verify the OTP sent to the user's email or phone.",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "vishvadattfreshcode@gmail.com",
                "description": "Email address of the user."
              },
              "otp": {
                "type": "string",
                "example": "123456",
                "description": "OTP received by the user."
              },
              "fcmToken": {
                "type": "string",
                "example": "vxvfsdhfowhrowrowhrlsdvxfd",
                "description": "FCM token for push notifications."
              },
              "deviceType": {
                "type": "string",
                "example": "ios",
                "description": "Type of device used by the user (e.g., ios, android)."
              },
              "phone": {
                "type": "string",
                "example": "any",
                "description": "Phone number of the user (optional)."
              },
              "provider": {
                "type": "string",
                "example": "any",
                "description": "Authentication provider name (e.g., Google, Facebook)."
              },
              "providerId": {
                "type": "string",
                "example": "any",
                "description": "ID provided by the authentication provider."
              }
            },
            "required": ["email", "otp", "fcmToken", "deviceType"]
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "OTP verified successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 1,
                  "description": "Status of the request. 1 indicates success."
                },
                "message": {
                  "type": "string",
                  "example": "Verify Successfully.",
                  "description": "Response message indicating the result of the request."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                      "description": "JWT token for authenticated user."
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "675aa7ff6fea1c0bd1881f07",
                          "description": "Unique ID of the user."
                        },
                        "email": {
                          "type": "string",
                          "example": "vishvadattfreshcode@gmail.com",
                          "description": "Email address of the user."
                        },
                        "userName": {
                          "type": "string",
                          "example": "vishu12",
                          "description": "Username of the user."
                        },
                        "isVerified": {
                          "type": "boolean",
                          "example": true,
                          "description": "Indicates if the user's email is verified."
                        },
                        "providerId": {
                          "type": "string",
                          "example": null,
                          "description": "ID of the authentication provider."
                        },
                        "providerName": {
                          "type": "string",
                          "example": "manually",
                          "description": "Name of the authentication provider."
                        },
                        "isOnline": {
                          "type": "boolean",
                          "example": false,
                          "description": "Indicates if the user is currently online."
                        },
                        "lastSeen": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-12-12T09:08:15.808Z",
                          "description": "Timestamp of the user's last online activity."
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-12-12T09:08:15.808Z",
                          "description": "Timestamp of user account creation."
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-12-12T09:08:15.808Z",
                          "description": "Timestamp of last user account update."
                        },
                        "__v": {
                          "type": "integer",
                          "example": 1,
                          "description": "Version key for the user document."
                        },
                        "otp": {
                          "type": "string",
                          "example": null,
                          "description": "The OTP value (cleared after verification)."
                        },
                        "otpExpiry": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-12-30T05:53:59.173Z",
                          "description": "Timestamp when the OTP expires."
                        },
                        "profilePicture": {
                          "type": "string",
                          "example": "images-Screenshot 2024-06-07 131757.png",
                          "description": "URL of the user's profile picture."
                        },
                        "clearedChats": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [],
                          "description": "List of chat IDs cleared by the user."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Bad Request - Invalid input or missing parameters."
      },
      "500": {
        "description": "Internal Server Error - An error occurred on the server."
      }
    }
  }
},
...
    "/v1/users": {
  "get": {
    "tags": ["Users"],
    "summary": "Get All Users",
    "description": "Retrieve a paginated list of users. Supports filtering by username.",
    "parameters": [
      {
        "name": "authorization",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Bearer token for authentication."
      },
      {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "example": 1
        },
        "description": "Page number for pagination."
      },
      {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "example": 10
        },
        "description": "Number of users per page for pagination."
      },
      {
        "name": "username",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "example": "dani1234"
        },
        "description": "Filter users by username."
      }
    ],
    "responses": {
      "200": {
        "description": "Users retrieved successfully.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 1,
                  "description": "Status of the request. 1 indicates success."
                },
                "message": {
                  "type": "string",
                  "example": "Get All Users.",
                  "description": "Response message indicating the result of the request."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "675aaa0a1c1ec3a1f9a8ea01",
                            "description": "Unique ID of the user."
                          },
                          "phone": {
                            "type": "string",
                            "example": "9723966918",
                            "description": "Phone number of the user."
                          },
                          "email": {
                            "type": "string",
                            "example": "user2@gmail.com",
                            "description": "Email address of the user."
                          },
                          "userName": {
                            "type": "string",
                            "example": "dani1234",
                            "description": "Username of the user."
                          },
                          "isVerified": {
                            "type": "boolean",
                            "example": true,
                            "description": "Indicates if the user's email or phone is verified."
                          },
                          "providerId": {
                            "type": "string",
                            "example": "123456789abcd",
                            "description": "ID of the authentication provider."
                          },
                          "providerName": {
                            "type": "string",
                            "example": "google",
                            "description": "Name of the authentication provider."
                          },
                          "isOnline": {
                            "type": "boolean",
                            "example": false,
                            "description": "Indicates if the user is currently online."
                          },
                          "lastSeen": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-12-12T09:16:58.656Z",
                            "description": "Timestamp of the user's last online activity."
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-12-12T09:16:58.656Z",
                            "description": "Timestamp of user account creation."
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-12-12T09:16:58.656Z",
                            "description": "Timestamp of last user account update."
                          },
                          "profilePicture": {
                            "type": "string",
                            "example": "/uploads/1734088117864-Scenario4.png",
                            "description": "URL of the user's profile picture."
                          }
                        }
                      },
                      "description": "List of user objects."
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 2,
                      "description": "Total number of pages available."
                    },
                    "currentPage": {
                      "type": "integer",
                      "example": 1,
                      "description": "Current page number."
                    },
                    "totalUsers": {
                      "type": "integer",
                      "example": 4,
                      "description": "Total number of users available."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized - Authentication required."
      },
      "403": {
        "description": "Forbidden - User does not have permission to perform this action."
      },
      "500": {
        "description": "Internal Server Error - An error occurred on the server."
      }
    }
  }
},
    "/v1/chat/create-chat": {
  "post": {
    "tags": ["Chat"],
    "summary": "Create New Chat",
    "description": "Creates a new one-to-one chat with another user.",
    "parameters": [
      {
        "name": "authorization",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Bearer token for authentication."
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "userId2": {
                "type": "string",
                "example": "67601dfa6307547d19944818",
                "description": "The ID of the user to initiate the chat with."
              }
            },
            "required": ["userId2"]
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Chat created successfully.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 1,
                  "description": "Status of the request. 1 indicates success."
                },
                "message": {
                  "type": "string",
                  "example": "New chat created.",
                  "description": "Response message indicating the result of the request."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "6763ec2315fdf3b3cdf3de56",
                      "description": "Unique ID of the newly created chat."
                    },
                    "isGroup": {
                      "type": "boolean",
                      "example": false,
                      "description": "Indicates if the chat is a group chat."
                    },
                    "participants": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "675aa7ff6fea1c0bd1881f07",
                        "67601dfa6307547d19944818"
                      ],
                      "description": "List of user IDs participating in the chat."
                    },
                    "admins": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [],
                      "description": "List of admin IDs in the chat. Empty for one-to-one chats."
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-12-19T09:49:23.755Z",
                      "description": "Timestamp when the chat was created."
                    },
                    "__v": {
                      "type": "integer",
                      "example": 0,
                      "description": "Version key maintained by Mongoose."
                    },
                    "lastMessage": {
                      "type": "string",
                      "example": "6763ef0a8c568c9605454aa0",
                      "description": "ID of the last message in the chat."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized - Authentication required."
      },
      "403": {
        "description": "Forbidden - User does not have permission to perform this action."
      },
      "500": {
        "description": "Internal Server Error - An error occurred on the server."
      }
    }
  }
},
...
    "/v1/chat/upload-media": {
      "post": {
        "tags": ["Upload"],
        "description": "Upload media on server.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/v1/push": {
      "post": {
        "tags": ["Push"],
        "description": "push notification",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "example": "any"
                  },
                  "deviceToken": {
                    "example": "any"
                  },
                  "deviceType": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}